claude_agent/requirements.txt
django>=5.0
djangorestframework>=3.14
python-dotenv>=1.0.0
anthropic>=0.46.0
python-docx>=0.8.11
PyPDF2>=3.0.0
openpyxl>=3.1.0
drf-yasg>=1.21.5 